// src/app/encuestas/interfaces/encuestas.interface.ts
export interface EncuestadoInterface {
    per_id: number;
    car_id: number;
    per_cedula: string;
    per_nombre: string;
    per_cargo: string;
    per_clave: string;
    per_linea: string;
    per_estado: string;
    creacion_usuario: string;
    creacion_fecha: Date;
    modifica_usuario: string;
    modifica_fecha: Date;
}

// Generated by https://quicktype.io

export interface AsignadosInterface {
    per_id?: number;
    per_id_asignado: number;
    enc_id: number;
    asi_estado: string;
    per_nombre: string;
    per_cargo: string;
    asig_nombre: string;
    asig_cargo: string;
}

// Generated by https://quicktype.io

export interface BloquesInterface {
    blo_id: number;
    enc_id: number;
    blo_nombre: string;
    blo_orden: number;
    preguntas: PreguntasInterface[];
}
export interface PreguntasInterface {
    pre_id: number;
    blo_id: number;
    enc_id: number;
    pre_nombre: string;
    respuestas: RespuestaInterface[];
    valor: RespuestaInterface;
}

export interface RespuestaInterface {
    res_id: number;
    pre_id: number;
    res_etiqueta: string;
    res_valor: number;
    res_orden: number;
}

// Generated by https://quicktype.io

export interface CargoInterface {
    car_id: number;
    car_nombre: string;
    car_estado: string;
    creacion_usuario: string;
    creacion_fecha: Date;
    modifica_usuario: string;
    modifica_fecha: Date;
    estado: string;
}

// Generated by https://quicktype.io

export interface GrupoEncuestaInterface {
    gru_id: number;
    gru_nombre: string;
    gru_estado: string;
    creacion_usuario: string;
    creacion_fecha: Date;
    modifica_usuario: string;
    modifica_fecha: Date;
    estado: string;
}

/**
 * desde aqui los cambios
 * */
export interface SeccionInterface {
    sec_id: number;
    sec_nombre: string;
    sec_uuid: string;
    preguntas: PreguntaInterface[];
}

export interface PreguntaInterface {
    pre_id: number;
    pre_nombre: string;
    pre_tipo: string;
    pre_uuid: string;
    opciones: OpcionInterface[];
}

export interface OpcionInterface {
    opc_id: number;
    opc_nombre: string;
    opc_uuid?: string;
    opc_valor?: boolean;
}

export interface FormularioInterface {
    enc_id: number;
    enc_grupo: number;
    enc_nombre: string;
    enc_estado: string;
    enc_descripcion: string;
    creacion_usuario: string;
    creacion_fecha: Date;
    modifica_usuario: string;
    modifica_fecha: Date;
    formulario: any;
}
export interface GrupoEncuestaInterface {
    gru_id: number;
    gru_nombre: string;
    gru_estado: string;
    gru_fecha_inicio: Date;
    gru_fecha_final: Date;
    creacion_usuario: string;
    creacion_fecha: Date;
    modifica_usuario: string;
    modifica_fecha: Date;
}

export interface PlantillaCompetenciasInterface {
    nombre_empleado: string | undefined;
    cedula: string;
    correo: string;
    cargo: string;
    competencias: string;
    nivel_de_competencia_requerido: string;
    departamento: string;
    area: string;
    evaluadores?: any [];
}

export interface CargoFormularioInterface {
    nombre_cargo: string;
    id_formulario: any;
    nombre_formulario: string;
}